renamed adapater parse to receive (only used for incoming)
