+++ lib/adapter/constructor.js
@@ -5,7 +5,7 @@ module.exports = new Observable({
     id: require('vigour-js/lib/util/uuid') // default gets .val
   },
   inject: [
-    require('./parse'),
+    require('./recieve'),
     require('./scope'),
     require('./protocol') // just use id so you can choose one
   ],
